producer/requirements.txt
requests
kafka-python
pandas